backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
django-filter==24.1
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
pillow==10.1.0
python-dateutil==2.8.2
celery==5.3.4
redis==5.0.1
requests==2.31.0